+++ CHANGELOG.md
@@ -86,4 +86,6 @@ September 12th, 2023
 
 - Home Index now checks to make sure the user is signed in with DropAuth (or a DropAuth compliant service), rather than just checking for a username.
 - Improved the reliability of the database organization process.
-- Fixed database organization.
+- Added descriptions to various input fields.
+- Moved the 'rainbow' theme to the end of the list on the configuration page.
+- Added container list tool